web2e - Copy/requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
Werkzeug==2.3.7
reportlab==4.0.4
python-dotenv==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.6.2
click==8.1.7
gunicorn==21.2.0
packaging==23.1
